Removed connection.h &c and added the struct to network.h
I did this because I had no other use for connection.c besides holding this structure.

The current plan is to have the clients establish their connection to the server and the server will add them to a game. Once there are two players in a game, the server will call a function that starts the game and sends the information between the two connections as needed. I think the server should call the function when there are two players waiting instead of the clients calling the function on their end. The clients are just sending data to the server and the server handles the logic.